requirements.txt
uvicorn~=0.27.1
fastapi~=0.109.2
python-dotenv==1.0.0
Jinja2==3.1.6
aiofiles==24.1.0
requests==2.32.3
pydantic-settings==2.9.1
apscheduler==3.10.4